addedTaxRTGS

+++ analyserModule/api/Helpers/processTransactions.py
@@ -3,7 +3,6 @@ import pandas as pd
 import re
 import json
 from ..models import monthWiseAnalytics
-
 import requests
 from django.conf import settings
 import os
@@ -16,7 +15,6 @@ def preprocessing(transactions):
     transactions = transactions.replace(np.nan, 0)  # replacing NaN with 0.0
     requiredColumnslist = ['Date', 'Particulars', 'Debit',
                            'Credit', 'Balance']  # Columns list we need
-    transactions = transactions[requiredColumnslist]
     return transactions
 
 
@@ -40,6 +38,7 @@ def processing(transactions, accountNumber, token=None):
     year = getStartDate(transactions).get('year')
     endMonth = getEndDate(transactions).get('month')
     endYear = getEndDate(transactions).get('year')
+    mergedTransactions = pd.DataFrame()
     while (1):
         if (year > endYear) or ((year == endYear) and (month > endMonth)):
             break
@@ -49,61 +48,30 @@ def processing(transactions, accountNumber, token=None):
                 queriedTransactions = (transactions['month'] == month) & (
                     transactions['year'] == year)
 
-                # update this to banking microservice, and process this only when it's successfully updated
+                currTransactions = transactions[queriedTransactions]
 
+                # update this to banking microservice, and process this only when it's successfully updated
                 if (token):
-                    currTransactions = transactions[queriedTransactions]
-                    currTransactions.to_csv("transactions.csv")
-                    response = requests.post(getattr(settings, "BANKING_MICROSERVICE", None) + "add_transactions",
-                                             data={
-                                                 "account_number": accountNumber},
-                                             headers={'Authorization': token},
-                                             files={"transactions": "transactions.csv"})
-                    print(response)
-                    os.remove("transactions.csv")
+                    if mergedTransactions.empty:
+                        mergedTransactions = currTransactions
+                    else:
+                        mergedTransactions = pd.concat(
+                            [mergedTransactions, currTransactions])
                 disjointList.append([month, year, currTransactions])
         month += 1
         if month == 12:
             month = 0
             year += 1
+    if token and not mergedTransactions.empty:
+        mergedTransactions.to_csv("transactions.csv")
+        with open("transactions.csv") as f:
+            response = requests.post(settings.BANKING_MICROSERVICE + "add_transactions/",
+                                     data={"account_number": accountNumber},
+                                     headers={'Authorization': token},
+                                     files={"transactions": f})
+            print(response.json(), response.status_code)
+        os.remove("transactions.csv")
     return disjointList
-  disjointList = []
-  month = getStartDate(transactions).get('month')
-  year = getStartDate(transactions).get('year')
-  endMonth = getEndDate(transactions).get('month')
-  endYear = getEndDate(transactions).get('year')
-  mergedTransactions = pd.DataFrame()
-  while(1) : 
-    if (year > endYear) or ((year == endYear) and (month > endMonth)): 
-      break
-    else : 
-      # check with backend if that exists, if it does't
-      if(len(monthWiseAnalytics.objects.filter(month = month, year=year, accountNumber=accountNumber).values())== 0):		
-        queriedTransactions = (transactions['month'] == month) & (transactions['year'] == year)
-
-        currTransactions = transactions[queriedTransactions]
-        
-        # update this to banking microservice, and process this only when it's successfully updated
-        if(token):  
-          if  mergedTransactions.empty : 
-              mergedTransactions = currTransactions
-          else : 
-              mergedTransactions = pd.concat([mergedTransactions, currTransactions])
-        disjointList.append([month, year, currTransactions])
-    month += 1
-    if month == 12: 
-      month = 0
-      year += 1
-  if token and not mergedTransactions.empty: 
-    mergedTransactions.to_csv("transactions.csv")
-    with open("transactions.csv") as f:
-      response = requests.post(settings.BANKING_MICROSERVICE + "add_transactions/", 
-                              data={"account_number": accountNumber}, 
-                              headers = { 'Authorization': token },
-                              files={"transactions": f})
-      print(response.json(), response.status_code)
-    os.remove("transactions.csv")
-  return disjointList
 
 
 def createSearchBase():
